Spawning Biomass (SB)

The total weight of all fish (both males and females) in the population that contribute to reproduction. Often conventionally defined as the biomass of all individuals beyond “age at first maturity” or “size at first maturity,” i.e. beyond the age or size class in which 50 percent of the individuals are mature; 2. The total biomass of fish of reproductive age during the breeding season of a stock. Most often used as a proxy for measuring egg production, the SSB depends on the abundance of the various age classes composing the stock and their past exploitation pattern, rate of growth, fishing and natural mortality rates, onset of sexual maturity, and environmental conditions.

SpawningBiomass()

Format

Examples

female spawning biomass

Rationale

Spawning biomass and spawning stock biomass have both been used historically, though the former is shorter without sacrificing clarity. For single-sex models, spawning biomass often pertains only to females but text should be specific, e.g., female spawning biomass. The alternative, spawning output, is only viable for species that exhibit multiple spawning events per fish within a time period and have pelagic eggs.

Alternatives

spawning stock biomass, spawning output (eggs), spawning stock output (eggs)

Range of possible values

0--Inf

Units

mt